+++ client/waatea2_client/lib/models/abonnement_model.dart
@@ -1,10 +1,12 @@
 class AbonnementModel {
   final int pk;
   final String name;
+  final String short;
 
-  AbonnementModel({required this.pk, required this.name});
+  AbonnementModel({required this.pk, required this.name, required this.short});
 
   factory AbonnementModel.fromJson(Map<String, dynamic> json) {
-    return AbonnementModel(pk: json['pk'], name: json['name']);
+    return AbonnementModel(
+        pk: json['pk'], name: json['name'], short: json['short']);
   }
 }