+++ client/waatea2_client/lib/models/team_model.dart
@@ -1,0 +1,21 @@
+class TeamModel {
+  final String id;
+  final String name;
+
+  TeamModel({
+    required this.id,
+    required this.name,
+  });
+
+  factory TeamModel.fromJson(Map<String, dynamic> json) {
+    return TeamModel(
+      id: json['id'],
+      name: json['name'],
+    );
+  }
+
+  Map<String, dynamic> toJson() => {
+        'id': id,
+        'name': name,
+      };
+}